requirements.txt
pandas==2.1.1
matplotlib.pyplot==3.8.0
numpy ==1.26.0
seaborn==0.12.2
streamlit==1.27.1